bug fix and different middle click logic
fixed a bug introduced by a previous commit where victory state was being checked incorrectly.
changed all the board logic to not use polymorphism (faster and easier to read).
got rid of a lot of stuff that wasn't being used (some was simply inlined into a single call site).
made a lot of stuff private that used to be public.